feat(code-gen): move validator generator to the common directory and split per group
