+++ packages/code-gen/src/internal-exports.js
@@ -13,4 +13,4 @@ export {
   getSortedKeysForType,
   getPrimaryKeyWithType,
 } from "./generator/sql/utils.js";
-export { validateCodeGenStructure } from "./generated/index.js";
+export { validateCodeGenStructure } from "./generated/codeGen/validators.js";
